# [Proposal] Force the inlining of StringBuilder.Append(char)