@jkotas 2019-10-22
I do not think that this is the right throughput/codesize trade-off to make.