@EgorBo 2019-10-22
@adamsitnik I do it this way:
1) Builld CoreCLR repo, ideally full build with tests and `Checked` config:
```
build.cmd -checked
```
2) Clone & build https://github.com/dotnet/jitutils
3) run
```
jit-diff diff --output C:\prj\jit-diff -f --pmi --core_root C:\prj\coreclr\bin\tests\Windows_NT.x64.Checked\Tests\Core_Root --base C:\prj\coreclr\bin\Product\Windows_NT.x64.Checked
```
Apply change, build corelib:
```
build.cmd -checked -skiptests -skipnative
```
4) run command from 3) again
5) run
```
jit-analyze.bat --base C:\prj\jit-diff\dasmset_1\base --diff C:\prj\jit-diff\dasmset_2\base\ --recursive -c 100
```
Should print a diff.