@danmosemsft 2019-10-22
>> the same way as List.Add does

> I think this change is still debatable.

I just noticed List<T>.Clear was force inlined. That was more surprising to me. I would have thought that was significantly less common. It also seems maybe larger than Add. https://github.com/dotnet/coreclr/pull/11104