@jkotas 2019-10-22
> List.Clear 

List.Clear has a large opportunity for dead-code removal. See the discussion in the issue that introduced this.

>  thought that was significantly less common.

That means the cost of aggressively inlining it is significantly less. Aggressive inlining is most problematic on methods that are used frequently.