@jkotas 2019-10-22
The string building methods are often big, with a mix of logic and Append calls. The problem with aggresively inlining code into such methods is that the JIT will hit the threshold where it gives up optimizations much faster and the method overall will get much slower. It should be easy to build a micro-benchmark based on pattern like this that shows a significant regression due to this change.