@jkotas 2019-10-22
It is unlikely that somebody will be calling `builder.Append` in a loop like the micro-benchmark. `builder.Append('a', length);` is both faster and more concise.