@JeremyKuhne 2019-10-22
I agree with @jkotas- I don't think we should inline this particular method.