@adamsitnik 2019-10-23
> It is unlikely that somebody will be calling builder.Append in a loop like the micro-benchmark

`StringBuilder.Append(char)` is frequently used in a loop that filters out some characters from the source string. An example:

https://github.com/dotnet/BenchmarkDotNet/blob/0768b7206be0078516034f0e6a1690b61561c55c/src/BenchmarkDotNet/Diagnosers/DisassemblyPrettifier.cs#L123-L126

I must admit that when I was creating this PR I forgot about the other use case - multiple calls to `Append(char)` from a single method:

https://github.com/Bunny83/SimpleJSON/blob/1945ecb12b54ab84f62f43609b82616ae16742e2/SimpleJSON.cs#L712-L725

In this case, inlining would decrease performance.

I agree that forcing the inlining of this method is a bad idea. Ideally, JIT should inline it only in loops that call this method once. Thank you for the productive discussion!